blockchain/requirements.txt
requests>=2.31.0
fastapi>=0.96.0
pydantic>=1.10.8
uvicorn>=0.22.0
gunicorn>=20.0.1
python-dotenv
loguru
black